test(database): add backup and restore it after each test
Also move setup test module and database to separate files
